Reorganized; Parallel Top-M Search using OpenMP critical area for queues

+++ include/definitions.h
@@ -13,7 +13,9 @@ namespace PANNS {
 using dataf = float;
 using distf = float;
 using idi = uint32_t;
-using edgei = uint32_t;
+using edgei = uint64_t;
+//using edgei = uint32_t;
+
 //using Candidate = std::pair<distf, idi>; // .first: distance; .second: vertex ID
 //using Candidate = std::tuple<dataf, idi, bool>; // 0: distance; 1: vertex ID; 2: flag is checked
